requirements.txt
crewai[tools]
langchain-google-genai
python-dotenv
fastapi
uvicorn[standard]
pydantic
google-search-results
sse-starlette
